Fix of login logic, little fixes in main screen. first step of character creation screen

+++ app/src/main/java/com/example/dndredactor/ui/mainScreen/MainViewModel.kt
@@ -2,49 +2,68 @@ package com.example.dndredactor.ui.mainScreen
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.dndredactor.data.AppConstants.loadError
-import com.example.dndredactor.data.model.Character
+import com.example.dndredactor.R
+import com.example.dndredactor.data.CustomException
+import com.example.dndredactor.data.repository.AuthRepository
 import com.example.dndredactor.data.repository.CharacterRepository
+import com.example.dndredactor.data.storage.TokenStorage
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
 class MainViewModel(
-    private val characterRepository: CharacterRepository
+    private val characterRepository: CharacterRepository,
+    private val authRepository: AuthRepository,
+    private val tokenStorage: TokenStorage
 ) : ViewModel() {
-    private val _uiState = MutableStateFlow<MainUiState>(MainUiState.Loading)
+    private val _uiState = MutableStateFlow<MainScreenUiState>(MainScreenUiState.Loading)
     val uiState = _uiState.asStateFlow()
 
     init {
-        loadCharacters()
     }
 
     private fun loadCharacters() {
         viewModelScope.launch {
-            _uiState.value = MainUiState.Loading
+            _uiState.value = MainScreenUiState.Loading
+
             characterRepository.getCharacters().onSuccess { characters ->
-                _uiState.value = MainUiState.Success(characters)
-            }.onFailure { exception ->
-                _uiState.value = MainUiState.Error(exception.message ?: loadError)
+                _uiState.value = MainScreenUiState.Success(characters)
+            }.onFailure { error ->
+
+                if (tokenStorage.getAccessToken().isNullOrBlank() && tokenStorage.getRefreshToken().isNullOrBlank()) {
+                    _uiState.value = MainScreenUiState.LoggedOut
+                    return@onFailure
+                }
+
+                val errorId = if (error is CustomException) {
+                    error.stringResId
+                } else {
+                    R.string.characterLoadError
+                }
+
+                _uiState.value = MainScreenUiState.Error(errorId)
             }
         }
     }
 
     fun deleteCharacter(id: String) {
         viewModelScope.launch {
-            val cur = (uiState.value as MainUiState.Success).characters.toMutableList()
-            cur.removeAll { it.id == id }
-            _uiState.value = MainUiState.Success(cur)
+            val current = uiState.value
+            if (current !is MainScreenUiState.Success) return@launch
+            val previous = current.characters
+            val newList = previous.toMutableList().apply { removeAll { it.id == id } }
+            _uiState.value = MainScreenUiState.Success(newList)
 
             characterRepository.deleteCharacter(id).onFailure {
-                loadCharacters()
+                _uiState.value = MainScreenUiState.Success(previous)
             }
         }
     }
-}
 
-sealed interface MainUiState {
-    object Loading : MainUiState
-    data class Success(val characters: List<Character>) : MainUiState
-    data class Error(val message: String) : MainUiState
+    fun onLogout() {
+        viewModelScope.launch {
+            authRepository.logout()
+            _uiState.value = MainScreenUiState.LoggedOut
+        }
+    }
 }
